+++ resources/views/test/index.blade.php
@@ -561,13 +561,32 @@
         },
 
         deleteTest(test){
-          axios.delete(window.Laravel.url+"/deletetest/"+test.id)
-          .then(response =>{
-            this.tests = response.data.tests;  
+
+          Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'danger',
+            showCancelButton: true,
+            confirmButtonColor: '#d33',
+            cancelButtonColor: '#ccc',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.value) {
+              Swal.fire(
+                'Deleted!',
+                '',
+                'success'
+              )
+              axios.delete(window.Laravel.url+"/deletetest/"+test.id)
+              .then(response =>{
+                this.tests = response.data.tests;  
+              })
+              .then(error =>{
+                console.log(error);
+              })
+            }
           })
-          .then(error =>{
-            console.log(error);
-          })
+          
         },
 
         moyenne(input){
